+++ context/app/static/js/components/savedLists/SavedListPanel/SavedListPanel.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Panel } from 'js/shared-styles/panels';
 
-import { useEntityCounts } from './hooks';
+import { useSavedEntityTypeCounts } from './hooks';
 
 function SavedListPanel({ entityObject, listUUID }) {
   const { savedEntities, description, title } = entityObject;
-  const counts = useEntityCounts(savedEntities);
+  const counts = useSavedEntityTypeCounts(savedEntities);
   return <Panel title={title} href={`/my-lists/${listUUID}`} secondaryText={description} entityCounts={counts} />;
 }
