+++ context/app/static/js/helpers/functions.js
@@ -98,3 +98,19 @@ export function addRestrictionsToQuery(query) {
     ...rest,
   };
 }
+
+export function getSearchHitsEntityCounts(searchHits) {
+  const counts = searchHits.reduce(
+    (acc, { _source: { entity_type } }) => {
+      if (!(entity_type in acc)) {
+        // Support entities may be in the user's list.
+        acc[entity_type] = 0;
+      }
+      const incrementedCount = acc[entity_type] + 1;
+      return { ...acc, [entity_type]: incrementedCount };
+    },
+    { Donor: 0, Sample: 0, Dataset: 0 },
+  );
+
+  return counts;
+}
